Reorder the consent page to be before selecting an ID to prepare for adding more information about the request.
