+++ src/components/LoginConsent/Consent/Consent.js
@@ -4,12 +4,8 @@ import { setExternalAction, setNavigationPath } from '../../../redux/reducers/na
 import { 
   ConsentRender
 } from './Consent.render';
-import { EXTERNAL_ACTION, EXTERNAL_CHAIN_START, REDIRECT, SCOPES, SELECT_LOGIN_ID } from '../../../utils/constants'
+import { EXTERNAL_ACTION, EXTERNAL_CHAIN_START, SCOPES, SELECT_LOGIN_ID } from '../../../utils/constants'
 import { checkAndUpdateIdentities } from '../../../redux/reducers/identity/identity.actions';
-import { signResponse } from '../../../rpc/calls/signResponse';
-import { setError } from '../../../redux/reducers/error/error.actions';
-import { LoginConsentDecision, LoginConsentResponse } from 'verus-typescript-primitives';
-import BigNumber from 'bignumber.js';
 
 class Consent extends React.Component {
   constructor(props) {
@@ -57,30 +53,7 @@ class Consent extends React.Component {
       userActions.map(action => this.props.dispatch(action))
 
       if (this.props.canLoginOrGiveConsent()) {
-        try {
-          let response = new LoginConsentResponse({
-            system_id: request.system_id,
-            signing_id: this.props.activeIdentity.identity.identityaddress,
-            decision: new LoginConsentDecision({
-              decision_id: request.challenge.challenge_id,
-              request: request,
-              created_at: BigNumber(Date.now())
-                .dividedBy(1000)
-                .decimalPlaces(0)
-                .toNumber(),
-            })
-          });
-
-          response.chain_id = request.chain_id
-          
-          const sigRes = await signResponse(response);
-          
-          this.props.setRequestResult(sigRes, () => {
-            this.props.dispatch(setNavigationPath(REDIRECT))
-          })
-        } catch(e) {
-          this.props.dispatch(setError(e))
-        }
+        this.props.dispatch(setNavigationPath(SELECT_LOGIN_ID));
       } else {
         this.props.dispatch(setExternalAction(EXTERNAL_CHAIN_START))
         this.props.dispatch(setNavigationPath(EXTERNAL_ACTION));
@@ -89,7 +62,9 @@ class Consent extends React.Component {
   }
 
   cancel() {
-    this.props.dispatch(setNavigationPath(SELECT_LOGIN_ID));
+    this.setState({ loading: true }, async () => {
+      await this.props.completeLoginConsent()
+    })
   }
 
   render() {
